requirements.txt
docker-py
python-consul
python-etcd
PyYAML==3.12
terminaltables>=3.1.0